+++ tests/testthat/test_odf.R
@@ -134,23 +134,41 @@ test_that("ODF header", {
 # https://github.com/dankelley/oce/issues/1801
 test_that("ODF temperature scale, IPTS68 and ITS90", {
           file <- system.file("extdata", "CTD_BCD2014666_008_1_DN.ODF.gz", package="oce")
-          # Read temperature from file (which is IPTS68) and convert to ITS90
+          # Get IPTS68 temperature, ITS90, and unit, all needed for tests.
+          # Note that all read() calls need to check for a warning that results
+          # from an error in this ODF file (and quite a few ODF files, it seems).
           Tref68 <- read.table(file, skip=675, header=FALSE)$V4
           Tref90 <- T90fromT68(Tref68)
+          unit <- list(unit=expression(degree*C), scale="IPTS-68")
           # read.odf()
           ctd1 <- expect_warning(read.odf(file), "should be unitless")
-          expect_equal(ctd1@metadata$units$temperature, list(unit=expression(degree*C), scale="IPTS-68"))
+          expect_equal(ctd1@metadata$units$temperature, unit)
           expect_equal(ctd1@data$temperature, Tref68)
           expect_equal(ctd1[["temperature"]], Tref90)
           # read.ctd.odf()
           ctd2 <- expect_warning(read.ctd.odf(file), "should be unitless")
-          expect_equal(ctd2@metadata$units$temperature, list(unit=expression(degree*C), scale="IPTS-68"))
+          expect_equal(ctd2@metadata$units$temperature, unit)
           expect_equal(ctd2@data$temperature, Tref68)
           expect_equal(ctd2[["temperature"]], Tref90)
           # read.oce()
           ctd3 <- expect_warning(read.oce(file), "should be unitless")
-          expect_equal(ctd3@metadata$units$temperature, list(unit=expression(degree*C), scale="IPTS-68"))
+          expect_equal(ctd3@metadata$units$temperature, unit)
           expect_equal(ctd3@data$temperature, Tref68)
           expect_equal(ctd3[["temperature"]], Tref90)
+          # as.ctd(read.odf())
+          ctd1B <- as.ctd(ctd1)
+          expect_equal(ctd1B@metadata$units$temperature, unit)
+          expect_equal(ctd1B@data$temperature, Tref68)
+          expect_equal(ctd1B[["temperature"]], Tref90)
+          # as.ctd(read.ctd.odf())
+          ctd2B <- as.ctd(ctd2)
+          expect_equal(ctd2B@metadata$units$temperature, unit)
+          expect_equal(ctd2B@data$temperature, Tref68)
+          expect_equal(ctd2B[["temperature"]], Tref90)
+          # as.ctd(read.oce())
+          ctd3B <- as.ctd(ctd3)
+          expect_equal(ctd3B@metadata$units$temperature, unit)
+          expect_equal(ctd3B@data$temperature, Tref68)
+          expect_equal(ctd3B[["temperature"]], Tref90)
 })